+++ Basil/include/Scene/Scene.h
@@ -17,10 +17,14 @@ namespace Basil
 			Scene();
 			~Scene();
 			Entity createEntity(const std::string& name = std::string());
+			void destroyEntity(Entity entity);
 			entt::registry& reg();
 			void onUpdate(Timestep timeStep);
 			void onViewportResize(uint32_t width, uint32_t height);
 		private:
+			template <typename T>
+			void onComponentAdded(Entity entity, T& component);
+
 			entt::registry registry;
 			uint32_t viewportWidth;
 			uint32_t viewportHeight;